move tus server implementation to strategies
