+++ apps/api/src/storage/src/strategy/strategy.ts
@@ -1,3 +1,4 @@
+import {StorageObjectMeta} from "@spica-server/interface/storage";
 import {ReadStream} from "fs";
 
 export abstract class Strategy {
@@ -7,6 +8,6 @@ export abstract class Strategy {
   abstract delete(id: string): Promise<void> | void;
   abstract url(id: string): Promise<string>;
   abstract rename(oldKey: string, newKey: string): Promise<void>;
-  abstract getTusServerDatastore(): any;
-  abstract getCleanUpExpiredUploadsMethod(): () => Promise<number> | null;
+  abstract handleResumableUpload(req: any, res: any): any;
+  abstract onResumableUploadFinished(callback: (document: StorageObjectMeta) => void): void;
 }
